More generic line parser and builder objects
Move the generic type arguments from the class level to the method level
for line parsers and builders so that multiple parser and builder objects
do not need to be created in order to work with multiple record types.

This is in pursuit of supporting an engine that can parse multiple record
types from the same file.

All tests pass, though it does introduce a breaking API change.