Fixed bug with ragged right files
If the length of a line in a file falls short of providing a value for
each field (ragged right file), the linePosition can end up equal to the
length of the line. This doesn't cause a problem if this is the last
field, but does pose a problem if there are additional fields to be
checked.

On getting the value for the next field the linePosition ends up after
the end of the line and throws an exception. The change makes sure that
the linePosition is either at or below the end of the line length before
trying to start the Substring function. This forces either the NullValue
to be returned, or the exception defined a few lines below to be thrown.